requirements.txt
# # attrs==24.2.0
# # blinker==1.8.2
# # certifi==2024.8.30
# # charset-normalizer==3.4.0
# # click==8.1.7
# # coloredlogs==15.0.1
# # cycler==0.12.1
# # filelock==3.16.1
# Flask==3.0.3
# flatbuffers==24.3.25
# fonttools==4.54.1
# fsspec==2024.9.0
# gunicorn==23.0.0
# h11==0.14.0
# humanfriendly==10.0
# idna==3.10
# itsdangerous==2.2.0
# Jinja2==3.1.4
# kiwisolver==1.4.7
# # MarkupSafe==3.0.0
# # matplotlib==3.9.2
# # mpmath==1.3.0
# # networkx==3.3
# # numpy==2.1.2
# numpy
# # onnxruntime==1.19.2
# opencv-python==4.10.0.84
# # outcome==1.3.0.post0
# # packaging==24.1
# pandas==2.2.3
# pillow==10.4.0
# # protobuf==5.28.2
# # psutil==6.0.0
# # py-cpuinfo==9.0.0
# # pyparsing==3.1.4
# # PySocks==1.7.1
# # python-dateutil==2.9.0.post0
# # pytz==2024.2
# # PyYAML==6.0.2
# # requests==2.32.3
# # scipy==1.14.1
# # seaborn==0.13.2
# # selenium==4.25.0
# # setuptools==75.1.0
# # six==1.16.0
# # sniffio==1.3.1
# # sortedcontainers==2.4.0
# # sympy==1.13.3
# torch==2.4.1
# torchvision==0.19.1
# tqdm==4.66.5
# # trio==0.26.2
# trio-websocket==0.11.1
# # typing_extensions==4.12.2
# # tzdata==2024.2
# ultralytics==8.3.8
# ultralytics-thop==2.0.9
# urllib3==2.2.3
# websocket-client==1.8.0
# Werkzeug==3.0.4
# # wheel==0.44.0
# # wsproto==1.2.0



Flask
torch
ultralytics
opencv-python
numpy
Pillow
Werkzeug
gunicorn

